requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cryptography==3.4.7
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.7
django-braces==1.14.0
django-ckeditor==6.0.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-js-asset==1.2.2
django-mptt==0.12.0
django-oauth-toolkit==1.5.0
django-taggit==1.3.0
django-taggit-serializer==0.1.7
djangorestframework==3.12.3
filelock==3.0.12
idna==2.10
isort==5.6.4
jwcrypto==0.8
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
psycopg2==2.8.6
pycparser==2.20
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
virtualenv==20.1.0
wrapt==1.12.1
